Moving utility methods about binary references to a separate file that can be accessed by the engine
