+++ engine/core/src/test/java/es/eucm/ead/engine/utils/ReferenceUtilsTest.java
@@ -34,7 +34,7 @@
  *      You should have received a copy of the GNU Lesser General Public License
  *      along with eAdventure.  If not, see <http://www.gnu.org/licenses/>.
  */
-package es.eucm.ead.editor.utils;
+package es.eucm.ead.engine.utils;
 
 import com.badlogic.gdx.utils.Array;
 import es.eucm.ead.schema.renderers.SpineAnimation;
@@ -49,10 +49,9 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
 /**
- * Created by jtorrente on 26/11/14.
+ * Created by jtorrente on 04/11/2015.
  */
-public class ProjectUtilsTest {
-
+public class ReferenceUtilsTest {
 	private static Array<String> buildUris() {
 		return new Array<String>(new String[] { "images/image1.png",
 				"images/image2.JPEG", "videos/video1.mpg", "video2.avi",
@@ -62,13 +61,13 @@ public class ProjectUtilsTest {
 
 	@Test
 	/**
-	 * Tests {@link es.eucm.ead.editor.utils.ProjectUtils#listRefBinaries(Object)}
+	 * Tests {@link es.eucm.ead.engine.utils.ReferenceUtils#listRefBinaries(Object)}
 	 */
 	public void testBinaryReferencesSearch() {
 		Array<String> uris = buildUris();
 		// Create a simple object
 		BinRefContainer binRefContainer = new BinRefContainer(uris);
-		Array<String> binaryReferences = ProjectUtils
+		Array<String> binaryReferences = ReferenceUtils
 				.listRefBinaries(binRefContainer);
 		for (int i = 0; i < uris.size; i++) {
 			String current = uris.get(i);
@@ -93,24 +92,24 @@ public class ProjectUtilsTest {
 
 	@Test
 	/**
-	 * Tests {@link es.eucm.ead.editor.utils.ProjectUtils#replaceBinaryRef(Object, String, String)}
+	 * Tests {@link es.eucm.ead.engine.utils.ReferenceUtils#replaceBinaryRef(Object, String, String)}
 	 */
 	public void testReplaceBinaryRef() {
 		Array<String> uris = buildUris();
 		// Create a simple object
 		BinRefContainer binRefContainer = new BinRefContainer(uris);
-		ProjectUtils
-				.replaceBinaryRef(binRefContainer, "A test", "Another test");
+		ReferenceUtils.replaceBinaryRef(binRefContainer, "A test",
+				"Another test");
 		assertEquals(binRefContainer.notAnUri, "Another test");
 		String str = "simple test";
-		ProjectUtils.replaceBinaryRef(binRefContainer, uris.get(2), str);
-		ProjectUtils.replaceBinaryRef(binRefContainer, uris.get(3), str);
-		ProjectUtils.replaceBinaryRef(binRefContainer, uris.get(4), str);
-		ProjectUtils.replaceBinaryRef(binRefContainer, uris.get(5), str);
-		ProjectUtils.replaceBinaryRef(binRefContainer, uris.get(6), str);
-		ProjectUtils
-				.replaceBinaryRef(binRefContainer, uris.get(8), "skeleton3");
-		ProjectUtils.replaceBinaryRef(binRefContainer, "skeleton3.json",
+		ReferenceUtils.replaceBinaryRef(binRefContainer, uris.get(2), str);
+		ReferenceUtils.replaceBinaryRef(binRefContainer, uris.get(3), str);
+		ReferenceUtils.replaceBinaryRef(binRefContainer, uris.get(4), str);
+		ReferenceUtils.replaceBinaryRef(binRefContainer, uris.get(5), str);
+		ReferenceUtils.replaceBinaryRef(binRefContainer, uris.get(6), str);
+		ReferenceUtils.replaceBinaryRef(binRefContainer, uris.get(8),
+				"skeleton3");
+		ReferenceUtils.replaceBinaryRef(binRefContainer, "skeleton3.json",
 				"skeleton2");
 		assertEquals(binRefContainer.refHolder.uri, str);
 		assertEquals(binRefContainer.list.get(0).uri, str);
@@ -167,4 +166,5 @@ public class ProjectUtilsTest {
 			this.uri = uri;
 		}
 	}
+
 }
